backends: Move Facebook Limited Login to separate module
This removes jose dependency from Facebook backend.

Fixes #773
